Cargo environment variables
This completes part 3 of project 1:
https://github.com/pingcap/talent-plan/blob/master/rust/projects/project-1/project.md#part-3-cargo-environment-variables

This replaces some hard-coded values into the cli.yml to instead
use the metadata stored in the Cargo.toml file at run-time. This means that
we no longer have multiple sources of truth, and multiple files to
maintain, since we can just use the environment variables Cargo sets
about the crate.